ckanext-ecportal/requirements-dev.txt
Babel==0.9.6
Beaker==1.6.4
FormAlchemy==1.4.3
FormEncode==1.2.6
Genshi==0.6
Jinja2==2.6
Mako==0.9.0
MarkupSafe==0.18
Pairtree==0.7.1-T
Paste==1.7.5.1
PasteDeploy==1.5.0
PasteScript==1.7.5
Pygments==1.6
Pylons==0.9.7
Routes==1.13
SPARQLWrapper==1.8.0
SQLAlchemy==0.8.7
Tempita==0.5.1
WebError==0.10.3
WebHelpers==1.3
WebOb==1.0.8
WebTest==1.4.3
apachemiddleware==0.1.1
argparse==1.2.1
decorator==3.4.0
fanstatic==0.12
html5lib==0.999
isodate==0.5.1
lxml==2.2.4
nose==1.3.0
ofs==0.4.1
ordereddict==1.1
paster==0.8
polib==1.0.3
psycopg2==2.4.5
pyparsing==2.0.3
python-dateutil==1.5
python-openid==2.2.5
pyutilib.component.core==4.5.3
rdflib==4.1.2
repoze.lru==0.6
repoze.who==2.1
repoze.who-friendlyform==1.0.8
repoze.who-use-beaker==0.3
repoze.who.plugins.openid==0.5.3
requests==2.1.0
simplejson==3.3.1
six==1.9.0
solrpy==0.9.5
sqlalchemy-migrate==0.7.2
unicodecsv==0.9.4
vdm==0.14
wsgiref==0.1.2
progressbar==2.3
kombu==2.5
celery==2.4.2
zope.interface==4.0.1
wheezy.validation==0.1.135
xmltodict==0.10.1
openpyxl==2.3.5
redis==2.10.5
ujson==1.35
datadiff==2.0.0
rfc3987==1.3.7
futures==3.2.0
tenacity==4.9.0